Créer une carte représentant les volumes en ronds proportionnels

Créer une carte représentant les volumes en ronds proportionnels

Usage

carte_rond_proportionnel(
  df,
  colonne_code_zone_geo,
  colonne_volume,
  titre_carte = "Carte 1",
  sous_titre_carte = "sous titre de la carte",
  titre_legende_volume = "titre de la legende",
  note_sous_carte = "Lecture : bla bla bla <br> <br> Source : Agreste",
  legend.position = "right",
  zone_geographique = "france",
  niveau_representation = "region_metro",
  couleur_fond_carte = "white",
  couleur_limite_zone = "black",
  taille_limite_zone = 0.3,
  couleur_fond_rond = "#E66A64",
  couleur_contour_rond = "darkgray",
  taille_max_rond = 9,
  hauteur_cm = 8,
  largeur_cm = 8,
  save_carte = FALSE,
  path_carte = "carte.png"
)

Arguments

df

dataframe avec les données à afficher sur la carte

colonne_code_zone_geo

nom de la colonne contenant les id de la zone géo : code région, n° de département...

colonne_volume

nom de la colonne contenant les données de volume

titre_carte

CHAR Titre à afficher en haut de la carte

sous_titre_carte

CHAR Sous-titre à afficher en haut de la carte, en dessous du titre

titre_legende_volume

CHAR Titre à afficher en haut de la légende

note_sous_carte

CHAR Note à afficher sous la carte

legend.position

CHAR choit de la position de la carte "right", "left", "bottom", "top"

zone_geographique

CHAR indiquant la zone retenue pour la carte : "france", "ue" ou "monde"

niveau_representation

CHAR indiquant le niveau de représentation : "pays", "region_metro" ou "departement_metro"

couleur_fond_carte

CHAR couleur de fond la carte "nom_couleur" ou "#code_couleur_html"

couleur_limite_zone

CHAR couleur des lignes de séparation de zone géographique "nom_couleur" ou "#code_couleur_html"

taille_limite_zone

NUM épaisseur des lignes de séparation de zone géographique

couleur_fond_rond

CHAR couleur de remplissage des ronds affichés "nom_couleur" ou "#code_couleur_html"

couleur_contour_rond

CHAR couleur de contour des ronds "nom_couleur" ou "#code_couleur_html"

taille_max_rond

NUM taille du rond correspondant au plus grand volume

hauteur_cm

NUM hauteur de la carte en cm

largeur_cm

NUM largeur de la carte en cm

save_carte

LGL TRUE si export de la carte au format png, FALSE par défaut (simple plot de la carte)

path_carte

CHAR chemin vers l'image à sauvegarder

Value

une carte de type ggplot 2

Examples

library(agreste)
library(dplyr)
### récupération des données de permis de construire de SITADEL 
permis_construire_region_2020 <- agreste::permis_construire_region %>% 
  filter(annee_autorisation == 2020)

### création de la carte et sauvegarde du fichier carte
carte_pc_2020 <- carte_rond_proportionnel(df = permis_construire_region_2020, colonne_code_zone_geo = REG, colonne_volume = PC, titre_carte = "Carte 1", sous_titre_carte = "Nombre de permis de construire délivré en 2020 par région", titre_legende_volume = "Nombre de permis de construire", save_carte = TRUE, path_carte = "carteReg.png")

### ouverture de la carte
browseURL("carteReg.png")

# Pour faire par département :
permis_construire_dep_2020 <- agreste::permis_construire_dep %>%
  filter(annee_autorisation == 2020)

carte_pc_2020 <- carte_rond_proportionnel(df = permis_construire_dep_2020, colonne_code_zone_geo = DEP, colonne_volume = PC, niveau_representation = "departement_metro", titre_carte = "Carte 2", sous_titre_carte = "Nombre de permis de construire délivré en 2020 par département", titre_legende_volume = "Nombre de permis de construire", save_carte = TRUE, path_carte = "carteDEP.png", taille_max_rond = 3)

browseURL("carteDEP.png")